Store extra bits and pieces as lists in the metadata.
If a particular scrapper step produces some miscellaneous output, this gets
stored in its own list, which is more compartmentalized than just adding each
individual object in the top-level metadata with a prefixed name.

We only need prefixes for the colData where it would be inconvenient to access
nested DataFrames. In such cases, the responsibility of prefixing the colnames
is moved out of the format* functions, which should only create the DF.